Import scalar types when unwrapping fields
This will add a map between types and where they're imported from. When
we later find a field that is unwrapped to an inner type we can see
where it comes from by looking in both the `class_dict` for other
generated classes or the `imported_types` map for custom imports such as
scalars.