Set type hints for parse and serialize to `typing.Any`
It's possible to use scalars for other types than `str`, f.ex. raw JSON
with objects and/or lists or floats.

Given a scalar type that's not a `pydantic` class we might want to
return a JSON serializable type like a `dict` or `list` from the
serialize function.

Similarly we might want to accept such type when parsing it to our
custom type.