Remove `Annotated` from return type
After landing #157 the field on the classes is now annotated with a
field call. When using the shorter results plugin we grab the whole
field (type and annotation) and use that as return type from the client.

Since we don't have support to add the `ast.Call`, the generated client
generates invalid code.

This PR will, instead of adding support for more expressions, filter out
`ast.Tuple` types which consists of an `ast.Subscript` followed by an
`ast.Call` where the called method is `Field`.